Do not auto-set 'tabstop'
'tabstop' is a *visual* setting, not concerned with any vagaries of a
project's indentation wishes.

In fact, one of the great reasons to insist on tab-based indenting is
that every developer can then pick their own tab width!

Sadly, the real world has intervened to make that option intractable,
but editorconfig-vim should nonetheless leave 'tabstop' alone.