Change line colorcolumn setting to not overwrite
This changes the modification of the 'colorcolumn' setting to not be
overwritten, but instead expanded when g:EditorConfig_max_line_indicator
is set to "line". This will preserve the previously set value. Also
instead of calculating the column ourselves, we let Vim do it, by adding
"+1" to the colorcolumn setting.