Set shiftwidth to 0 (defaults to tabstop value)
The current behavior is to set both shiftwidth and tabstop/softtabstop.
shiftwidth gets set to the same value as tabstop. If I open a file that
editorconfig indents with 8-space hard tabs, then ":set tabstop=4", my
shiftwidth will still be 8, meaning that when I indent I will get two
tabs. Set shiftwidth to 0, which defaults to the value of tabstop, and
doesn't set softtabstop either. This should have the same end result
with less complication.