Allow disabling softtabstop via g:EditorConfig_softtabstop_{space,tab} config
The 'softtabstop' setting is meant to be used with a value that is different
from the 'tabstop' setting (and 'tabstop' typically then kept at the default
8), and then renders the indent first with hard tabs, filling remaining
differences with spaces. Though this has some nice editing characteristics
(no messing with tabstop values but still getting cursor progressing to
certain columns), it is quite specific to Vim and hasn't caught on elsewhere.
Therefore, the EditorConfig spec doesn't support this feature.

The existing implementation sets 'softtabstop' to the 'tabstop' / indent
size, effectively turning off the mixed indenting completely. That is not wrong,
and actually necessary so that backspace will delete one complete indent instead
of a single space character when using spaces for indent.

However, when hard tabs are configured and the user (maybe temporarily) manually
adjusts the size of an indent by changing the 'tabstop' (and 'shiftwidth')
values, suddenly, 'tabstop' is different from 'softtabstop', and Vim starts
inserting the mix of tabs and spaces that the softtabstop feature is about!
Since Vim 7.3.693, such discrepancy can be avoided via a special negative value.
Also, some plugins or statuslines may wrongly interpret the positive
'softtabstop' value to indicate activation of the softtabstop feature, even
though it's effectively off.

This change is related to #137; however, instead of dropping the modification
of 'softtabstop' altogether, I think it's better to offer configuration options,
separately for tab and space indents, as the effect is quite different.
The default value of 1 maintains the existing behavior of setting 'softtabstop'
equal to 'shiftwidth'; 0 will turn off 'softtabstop' (not recommended with
spaces due to the strange backspace behavior, but in my opinion quite sensible
for tab-based indents).