Allow EditorConfig not touching softtabstop via [] config
Picking up the argument from @chreekat in #137, it can be argued that
'softtabstop' purely is a user experience feature and the EditorConfig plugin
should not modify it at all. (This argument can be made since Vim 7.3.693, where
a negative value makes 'softtabstop' automatically assume the value of
'shiftwidth'; a value of 0 disables the feature altogether. However, I think
most users would rather see 'softtabstop' as one of the indent options,
especially when positive values are assigned to it.)

A special (empty) List configuration value can make the EditorConfig plugin skip
the update of 'softtabstop'.