charset: if we changed fenc from '' to enc, set nomodified
If the buffer had no fileencoding, and we changed fenc to match the Vim-wide
encoding, we haven't actually changed the file.  However, we have set
modified on the buffer by virtue of assigning fenc.  In this case, set
nomodified since we didn't actually change anything.

This affects new buffers created during the VimEnter autocmd.