doc: move license, add header line
Move the license to the bottom, adding an additional help tag for it, so
that when the help page is initially opened, one can directly read the
table of content.

While at it, also add a short description to the first line of the help
page, so that it gets correctly displayed at :h local-additions.

Signed-off-by: Christian Brabandt <cb@256bit.org>